src/main/java/guru/springframework/spring5recipeapp/converters/CategoryCommandToCategory.java
package guru.springframework.spring5recipeapp.converters;


import guru.springframework.spring5recipeapp.commands.CategoryCommand;
import guru.springframework.spring5recipeapp.domain.Category;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

// converters are generally used to convert between domain types and command type
// as it is converter of CategoryCommand to Category
@Component // all converters are components
public class CategoryCommandToCategory implements Converter<CategoryCommand, Category>{

    @Synchronized
    @Nullable
    @Override
    public Category convert(CategoryCommand source) {
        if (source == null) {
            return null;
        }

        // as spring does not guarantee thread safety thus made it as final
        final Category category = new Category();
        category.setId(source.getId());
        category.setDescription(source.getDescription());
        return category;
    }
}
